Combined ShareActivity with FileDetailSharingFragment as much as possible
