+++ scripts/analysis/lint-results.txt
@@ -1,2 +1,2 @@
 DO NOT TOUCH; GENERATED BY DRONE
-      <span class="mdl-layout-title">Lint Report: 92 warnings</span>
+      <span class="mdl-layout-title">Lint Report: 91 warnings</span>
