+++ src/main/res/values/strings.xml
@@ -473,13 +473,7 @@
 
     <string name="share_dialog_title">Sharing</string>
     <string name="share_file">Share %1$s</string>
-    <string name="share_with_user_section_title">Share with users and groups</string>
     <string name="share_no_users">No data shared with users yet</string>
-    <string name="share_add_user_or_group">Add user or group</string>
-    <string name="share_via_link_edit_permission_label">Allow editing</string>
-    <string name="share_via_link_password_label">Protect with password</string>
-    <string name="share_via_link_password_title">Secured</string>
-    <string name="share_via_link_expiration_date_label">Set expiration date</string>
     <string name="share_via_link_menu_password_label">Password protect (%1$s)</string>
     <string name="share_expiration_date_label">Expires %1$s</string>
     <string name="share_no_expiration_date_label">Set expiration date</string>
@@ -489,7 +483,6 @@
     <string name="share_no_password_title">Set password</string>
     <string name="edit_permission_label">edit</string>
     <string name="share_via_link_hide_file_listing_permission_label">Hide file listing</string>
-    <string name="share_get_public_link_button">Get link</string>
     <string name="share_with_title">Share with…</string>
     <string name="share_with_edit_title">Share with %1$s</string>
     <string name="share_via_link_unset_password">Unset</string>
